+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/apis/BaseHubdleDelegateExtension.kt
@@ -1,8 +1,0 @@
-package com.javiersc.hubdle.project.extensions.apis
-
-import org.gradle.api.Project
-
-public interface BaseHubdleDelegateExtension : BaseHubdleExtension {
-
-    public val project: Project
-}