+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/apis/BaseHubdleExtension.kt
@@ -1,3 +1,8 @@
 package com.javiersc.hubdle.project.extensions.apis
 
-public interface BaseHubdleExtension
+import org.gradle.api.Project
+
+public interface BaseHubdleExtension {
+
+    public val project: Project
+}