+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/features/shared/HubdleKotlinComposeFeatureExtension.kt
@@ -8,7 +8,7 @@ import com.javiersc.hubdle.project.extensions._internal.PluginId
 import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
 import com.javiersc.hubdle.project.extensions._internal.library
 import com.javiersc.hubdle.project.extensions.android._internal.configureAndroidCommonExtension
-import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
@@ -87,7 +87,7 @@ constructor(
     }
 }
 
-public interface HubdleKotlinComposeDelegateFeatureExtension : BaseHubdleDelegateExtension {
+public interface HubdleKotlinComposeDelegateFeatureExtension : BaseHubdleExtension {
 
     public val compose: HubdleKotlinComposeFeatureExtension
         get() = project.getHubdleExtension()