+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/features/shared/HubdleKotlinExtendedStdlibFeatureExtension.kt
@@ -11,7 +11,7 @@ import com.javiersc.hubdle.project.extensions._internal.TEST_FUNCTIONAL
 import com.javiersc.hubdle.project.extensions._internal.TEST_INTEGRATION
 import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
 import com.javiersc.hubdle.project.extensions._internal.library
-import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
@@ -65,7 +65,7 @@ constructor(
     }
 }
 
-public interface HubdleKotlinExtendedStdlibDelegateFeatureExtension : BaseHubdleDelegateExtension {
+public interface HubdleKotlinExtendedStdlibDelegateFeatureExtension : BaseHubdleExtension {
 
     public val extendedStdlib: HubdleKotlinExtendedStdlibFeatureExtension
         get() = project.getHubdleExtension()