+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/features/shared/HubdleKotlinMoleculeFeatureExtension.kt
@@ -8,7 +8,7 @@ import com.javiersc.hubdle.project.extensions._internal.MAIN
 import com.javiersc.hubdle.project.extensions._internal.PluginId
 import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
 import com.javiersc.hubdle.project.extensions._internal.library
-import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
@@ -55,7 +55,7 @@ constructor(
     }
 }
 
-public interface HubdleKotlinMoleculeDelegateFeatureExtension : BaseHubdleDelegateExtension {
+public interface HubdleKotlinMoleculeDelegateFeatureExtension : BaseHubdleExtension {
 
     public val molecule: HubdleKotlinMoleculeFeatureExtension
         get() = project.getHubdleExtension()