+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/jvm/features/HubdleKotlinCompilerPluginFeatureExtension.kt
@@ -7,7 +7,7 @@ import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
 import com.javiersc.hubdle.project.extensions._internal.library
 import com.javiersc.hubdle.project.extensions._internal.libraryModule
 import com.javiersc.hubdle.project.extensions._internal.libraryPlatform
-import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
@@ -289,7 +289,7 @@ public enum class KotlinCompilerTestType(internal val dir: String) {
     Diagnostics("diagnostics"),
 }
 
-public interface HubdleKotlinCompilerPluginDelegateFeatureExtension : BaseHubdleDelegateExtension {
+public interface HubdleKotlinCompilerPluginDelegateFeatureExtension : BaseHubdleExtension {
 
     public val compiler: HubdleKotlinCompilerPluginFeatureExtension
         get() = project.getHubdleExtension()