+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/shared/features/HubdleIntellijFeatureExtension.kt
@@ -2,7 +2,7 @@ package com.javiersc.hubdle.project.extensions.shared.features
 
 import com.javiersc.hubdle.project.extensions.HubdleDslMarker
 import com.javiersc.hubdle.project.extensions._internal.getHubdleExtension
-import com.javiersc.hubdle.project.extensions.apis.BaseHubdleDelegateExtension
+import com.javiersc.hubdle.project.extensions.apis.BaseHubdleExtension
 import com.javiersc.hubdle.project.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.project.extensions.apis.enableAndExecute
 import com.javiersc.hubdle.project.extensions.java.hubdleJava
@@ -33,7 +33,7 @@ constructor(
     }
 }
 
-public interface HubdleIntellijDelegateFeatureExtension : BaseHubdleDelegateExtension {
+public interface HubdleIntellijDelegateFeatureExtension : BaseHubdleExtension {
 
     public val intellij: HubdleIntellijFeatureExtension
         get() = project.getHubdleExtension()